TDP-482: Replace Spark statistics computation by their match in Java 8. Removes all content store method that filters out TDP_ID (moves filtering to statistics operations). Removes Spark configuration from default configuration in dataset. Minor type changes (int -> long) to ensure no loss of precision when moving DQ's values to Data Prep.

+++ dataprep-backend-common/src/main/java/org/talend/dataprep/api/dataset/statistics/PatternFrequency.java
@@ -8,9 +8,9 @@ public class PatternFrequency {
     String pattern;
 
     @JsonProperty("occurrences")
-    int occurrences;
+    long occurrences;
 
-    public PatternFrequency(String pattern, int occurrences) {
+    public PatternFrequency(String pattern, long occurrences) {
         this.pattern = pattern;
         this.occurrences = occurrences;
     }
@@ -23,11 +23,11 @@ public class PatternFrequency {
         this.pattern = pattern;
     }
 
-    public int getOccurrences() {
+    public long getOccurrences() {
         return occurrences;
     }
 
-    public void setOccurrences(int occurrences) {
+    public void setOccurrences(long occurrences) {
         this.occurrences = occurrences;
     }
 }
